correcao de erros

+++ pages/listarFichaDeInternacao.php
@@ -1,4 +1,4 @@
-<h3 style="margin:15px;">Ficha de Internacao</h3>
+<h3 style="margin:15px;">Ficha de Internação</h3>
 <hr />
 <div class="row" style="padding:15px;">
     <div class="col-5">
@@ -17,8 +17,6 @@
     <tr>
         <th>ID</th>
         <th>NOME</th>
-        <th>NOME DO RESPONSAVEL</th> 
-        <th>MEDICO</th>
         <th>DATA DA INTERNAÇÃO</th>
         <th>JUSTIFICATIVA DA INTERNAÇÃO</th>
                   
@@ -26,26 +24,34 @@
         <th class="text-center">EXCLUIR</th>
     </tr>
 
-    <?php foreach ($FichadeInternacao as $FichadeInternacao) { ?>
-        <tr>
-            <td><?php echo $FichadeInternacao->id; ?></td>
-            <td><?php echo $FichadeInternacao->nome; ?></td>
-            <td><?php echo $FichadeInternacao->nomeDoResponsavel; ?></td>
-            <td><?php echo $FichadeInternacao->medico; ?></td>
-            <td><?php echo $FichadeInternacao->datadaInternaçao; ?></td>
-            <td><?php echo $FichadeInternacao->justificativadaInternaçao; ?></td>
 
 
-            <td class="text-center">
-            <a href="?page=FichadeInternacaoControle&acao=get&id=<?php echo $FichadeInternacao->id; ?>" class="btn btn-warning">
-                    <i class="bi bi-pencil"></i>
-                </a>
-            </td>
-            <td class="text-center">
-            <a href="?page=FichadeInternacaoControle&acao=excluir&id=<?php echo $FichadeInternacao->id; ?>" class="btn btn-danger">
-                    <i class="bi bi-trash"></i>
-                </a>
-            </td>
-        </tr>
-    <?php } ?>  
-</table>
+    <?php
+if (!empty($FichadeInternacoes) && (is_array($FichadeInternacoes) || is_object($FichadeInternacoes))) {
+?>
+    <table class="table">
+        <?php foreach ($FichadeInternacoes as $FichadeInternacao) { ?>
+            <tr>
+                <td><?php echo $FichadeInternacao->id; ?></td>
+                <td><?php echo $FichadeInternacao->nome; ?></td>
+                <td><?php echo $FichadeInternacao->medico; ?></td>
+                <td><?php echo date('d/m/Y', strtotime($FichadeInternacao->datadaInternaçao)); ?></td>
+                <td><?php echo $FichadeInternacao->justificativadaInternaçao; ?></td>
+
+                <td class="text-center">
+                    <a href="?page=FichadeInternacaoControle&acao=get&id=<?php echo $FichadeInternacao->id; ?>" class="btn btn-warning">
+                        <i class="bi bi-pencil"></i>
+                    </a>
+                </td>
+                <td class="text-center">
+                    <a href="?page=FichadeInternacaoControle&acao=excluir&id=<?php echo $FichadeInternacao->id; ?>" class="btn btn-danger">
+                        <i class="bi bi-trash"></i>
+                    </a>
+                </td>
+            </tr>
+        <?php } ?>
+    </table>
+<?php
+}
+?>
+
